NetBSD re-synchronization of the source tree
This brings our tree to NetBSD 7.0, as found on -current on the
10-10-2015.

This updates:
 - LLVM to 3.6.1
 - GCC to GCC 5.1
 - Replace minix/commands/zdump with usr.bin/zdump
 - external/bsd/libelf has moved to /external/bsd/elftoolchain/
 - Import ctwm
 - Drop sprintf from libminc

+++ minix/tests/test73.c
@@ -45,6 +45,13 @@ main(int argc, char *argv[])
 
 	start(73);
 
+	setuid(geteuid());
+
+	if(getuid() != 0) {
+		printf("Test 73 has to be run as root; test aborted\n");
+		exit(1);
+	}
+
 	unlink(pipefn);
 
 	/* 'big' as a substring indicates to testvm that it's ok to
